+++ stay__home/pubspec.yaml
@@ -24,6 +24,8 @@ dependencies:
   flutter:
     sdk: flutter
   geolocator: ^6.1.13
+  workmanager: ^0.2.0
+  flutter_local_notifications: ^3.0.3
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
